# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: AlecRae API
  description: |
    The AlecRae API provides programmatic access to the AlecRae email
    infrastructure platform. Send transactional and bulk emails, manage
    domains, configure webhooks, and query analytics.
  version: 0.1.0
  contact:
    name: AlecRae Support
    url: https://alecrae.com/support
    email: support@alecrae.com
  license:
    name: Proprietary
    url: https://alecrae.dev/terms
servers:
  - url: https://api.alecrae.com
    description: Production
  - url: https://api.staging.alecrae.com
    description: Staging
  - url: http://localhost:3001
    description: Local development
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /v1/messages:
    post:
      operationId: sendMessage
      summary: Send an email
      description: |
        Enqueues an email for delivery. Returns immediately with a message ID
        and a `202 Accepted` status. Use the message ID to poll for delivery
        status or configure webhooks for async notifications.
      tags: [Messages]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageInput"
      responses:
        "202":
          description: Message accepted for delivery
          headers:
            X-Request-Id:
              $ref: "#/components/headers/X-Request-Id"
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
          content:
            application/json:
              schema:
                type: object
                required: [id, status, createdAt]
                properties:
                  id:
                    type: string
                    description: Unique message identifier
                    example: a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4
                  status:
                    type: string
                    enum: [queued, sending]
                    example: sending
                  createdAt:
                    type: string
                    format: date-time
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"

    get:
      operationId: listMessages
      summary: List messages
      description: |
        Returns a paginated list of messages. Supports cursor-based pagination
        and filtering by status or tag.
      tags: [Messages]
      parameters:
        - $ref: "#/components/parameters/CursorParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: status
          in: query
          schema:
            type: string
            enum: [queued, sending, delivered, bounced, deferred, complained, failed]
        - name: tag
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated list of messages
          headers:
            X-Request-Id:
              $ref: "#/components/headers/X-Request-Id"
          content:
            application/json:
              schema:
                type: object
                required: [data, hasMore]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/MessageRecord"
                  cursor:
                    type: string
                    nullable: true
                    description: Cursor for the next page, or null if no more results
                  hasMore:
                    type: boolean
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/domains:
    post:
      operationId: addDomain
      summary: Add a domain
      description: |
        Register a new sending domain. Returns the required DNS records that
        must be configured before verification.
      tags: [Domains]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddDomainInput"
      responses:
        "201":
          description: Domain added successfully
          content:
            application/json:
              schema:
                type: object
                required: [data, message]
                properties:
                  data:
                    $ref: "#/components/schemas/DomainRecord"
                  message:
                    type: string
                    example: "Domain added. Configure the DNS records below, then verify."
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Domain already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        Pass a Bearer token in the Authorization header.
        Example: `Authorization: Bearer em_live_abc123...`
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Pass an API key in the X-API-Key header.
        Example: `X-API-Key: em_live_abc123...`
  headers:
    X-Request-Id:
      description: Unique identifier for the request, useful for support and debugging
      schema:
        type: string
        example: req_a1b2c3d4
    X-RateLimit-Limit:
      description: Maximum number of requests allowed in the current window
      schema:
        type: integer
        example: 1000
    X-RateLimit-Remaining:
      description: Number of requests remaining in the current window
      schema:
        type: integer
        example: 997
    X-RateLimit-Reset:
      description: Unix timestamp (seconds) when the rate limit window resets
      schema:
        type: integer
        example: 1712150400
  parameters:
    CursorParam:
      name: cursor
      in: query
      description: Pagination cursor from a previous response
      schema:
        type: string
    LimitParam:
      name: limit
      in: query
      description: Maximum number of items to return (1-100, default 20)
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
    SendMessageInput:
      type: object
      required: [from, to, subject]
      properties:
        from:
          type: string
          format: email
          description: Sender email address
          example: noreply@example.com
        to:
          type: array
          items:
            type: string
            format: email
          minItems: 1
          maxItems: 50
          description: Recipient email addresses
          example: ["user@example.com"]
        cc:
          type: array
          items:
            type: string
            format: email
          description: CC recipients
        bcc:
          type: array
          items:
            type: string
            format: email
          description: BCC recipients
        subject:
          type: string
          maxLength: 998
          description: Email subject line
          example: Welcome to AlecRae
        textBody:
          type: string
          description: Plain text body
        htmlBody:
          type: string
          description: HTML body
        tags:
          type: array
          items:
            type: string
          description: Tags for categorization and filtering
        metadata:
          type: object
          additionalProperties:
            type: string
          description: Custom key-value metadata (max 20 entries)
        scheduledAt:
          type: string
          format: date-time
          description: Schedule delivery for a future time
    MessageRecord:
      type: object
      required: [id, from, to, subject, status, createdAt, updatedAt]
      properties:
        id:
          type: string
        from:
          type: string
          format: email
        to:
          type: array
          items:
            type: string
            format: email
        cc:
          type: array
          items:
            type: string
            format: email
        bcc:
          type: array
          items:
            type: string
            format: email
        subject:
          type: string
        status:
          type: string
          enum: [queued, sending, delivered, bounced, deferred, complained, failed]
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deliveredAt:
          type: string
          format: date-time
          nullable: true
    AddDomainInput:
      type: object
      required: [domain]
      properties:
        domain:
          type: string
          description: The domain name to register
          example: example.com
    DomainRecord:
      type: object
      required: [id, domain, status, dnsRecords, spfVerified, dkimVerified, dmarcVerified, mxVerified, createdAt]
      properties:
        id:
          type: string
        domain:
          type: string
        status:
          type: string
          enum: [pending, verified, failed]
        dnsRecords:
          type: array
          items:
            $ref: "#/components/schemas/DnsRecord"
        spfVerified:
          type: boolean
        dkimVerified:
          type: boolean
        dmarcVerified:
          type: boolean
        mxVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
    DnsRecord:
      type: object
      required: [type, host, value]
      properties:
        type:
          type: string
          enum: [TXT, CNAME, MX, A, AAAA]
        host:
          type: string
        value:
          type: string
        priority:
          type: integer
          description: MX record priority
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [type, message, code]
          properties:
            type:
              type: string
              description: Error category
              example: validation_error
            message:
              type: string
              description: Human-readable error message
              example: "Invalid email address in 'to' field"
            code:
              type: string
              description: Machine-readable error code
              example: invalid_recipient
            context:
              type: object
              additionalProperties: true
              description: Additional error context
  responses:
    BadRequest:
      description: The request body or parameters are malformed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication credentials are missing or invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Authenticated but insufficient permissions for this operation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ValidationError:
      description: The request failed validation (invalid email, domain, etc.)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
        X-RateLimit-Limit:
          $ref: "#/components/headers/X-RateLimit-Limit"
        X-RateLimit-Remaining:
          $ref: "#/components/headers/X-RateLimit-Remaining"
        X-RateLimit-Reset:
          $ref: "#/components/headers/X-RateLimit-Reset"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: An unexpected server error occurred
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"